A copy
This is the shape that the json file is going to take shape. I am going to have this repository only contain this json file, so it can be hosted on its own. That way I am able to utilize the json file in order to plot the parks/ public spaces where people can go to, on my map.